Package the checks as an OCI container for usage in Kubernetes (#6)
* Add packaging OCI container

This commit adds a OCI to deploy the checks via Kubernetes

* Instruct OBS to build the container

This commit adds supporting files for building the container via OBS

* Publish container to OBS via GitHub actions

This commit expands the ci.yaml file wit the ability to deploy the
checks OCI image to OBS.

* Prohibit accidental overrides

This commit ensures checks get only installed into a non empty
pre-existing directory. This is done to ensure that the volume is
mounted and empty. The installation should fail otherwise.

* Remove dead code